Lab7-CNN/Code/requirements.txt
matplotlib==3.10.3
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
